Add feature user authentications and authorization

+++ backend/routes/posts.js
@@ -5,6 +5,8 @@ const Post = require("../models/post");
 
 const router = express.Router();
 
+const checkAuth = require("../middleware/middleware");
+
 const MIME_TYPE_MAP = {
   "image/png": "png",
   "image/jpeg": "jpg",
@@ -32,12 +34,13 @@ const storage = multer.diskStorage({
 });
 
 
-router.post("", multer({storage: storage}).single("image") ,(req, res, next) => {
+router.post("", checkAuth, multer({storage: storage}).single("image") ,(req, res, next) => {
   const url = req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url +"/images/" + req.file.filename
+    imagePath: url +"/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -50,7 +53,7 @@ router.post("", multer({storage: storage}).single("image") ,(req, res, next) => 
   });
 });
 
-router.put("/:id", multer({storage: storage}).single("image") , (req, res, next) => {
+router.put("/:id",checkAuth, multer({storage: storage}).single("image") , (req, res, next) => {
   let imagePath = req.body.imagePath;
   if(req.file){
     const url = req.protocol + '://' + req.get("host");
@@ -60,10 +63,15 @@ router.put("/:id", multer({storage: storage}).single("image") , (req, res, next)
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath:imagePath
+    imagePath:imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: "Update successful!" });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.modifiedCount > 0 ) {
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized" });
+    }
   });
 });
 
@@ -87,7 +95,7 @@ router.get("", (req, res, next) => {
   });
 });
 
-router.get("/:id", (req, res, next) => {
+router.get("/:id", checkAuth, (req, res, next) => {
   Post.findById(req.params.id).then(post => {
     if (post) {
       res.status(200).json(post);
@@ -97,10 +105,14 @@ router.get("/:id", (req, res, next) => {
   });
 });
 
-router.delete("/:id", (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+router.delete("/:id", checkAuth,(req, res, next) => {
+  Post.deleteOne({ _id: req.params.id , creator: req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if (result.deletedCount > 0 ) {
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized" });
+    }
   });
 });
